Handle sonar cloud coverage issue by writing more test
